made abstime monotonic, fixed entry 1 problem
This is in response to a bug Patrice found. The bug
he didn't find was the one that causes entry 1 to be ignored
if saber effs up and doesn't print the FFFF line indicating
a new file before the TRIG_ON line. episodes.py now waits
to reject entries on that basis until after the TRIG_OFF

The bug he reported had to do with stimuli showing up in episodes
where they were clearly not played. This resulted when saber
was stopped and started, which resets the abstime counter
and thus gives stimuli timestamps that cause them to be sorted
with episodes that took place in the distant past or future. I now
keep track of start events and update an offset variable that keeps
the abstime function monotonic

darcs-hash:20060927165244-21d25-eb4fba80a592b64d5c6e7dd0e3e25aa7ee0ec719.gz